requirements.txt
telethon
python-dotenv
pyyaml
